requirements.txt
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.12
segmentation-models-pytorch>=0.3.3
albumentations>=1.3.1
opencv-python-headless>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
openpyxl>=3.1.0
scikit-learn>=1.3.0
tqdm>=4.66.0
Pillow>=10.0.0
gradio>=4.44.1
